Move validation to service

+++ src/main/java/comflower/sagongsa/service/CommentService.java
@@ -1,11 +1,12 @@
 package comflower.sagongsa.service;
 
+import comflower.sagongsa.entity.Comment;
+import comflower.sagongsa.error.CommentNotFoundException;
+import comflower.sagongsa.error.PostNotFoundException;
+import comflower.sagongsa.repository.CommentRepository;
+import comflower.sagongsa.repository.PostRepository;
 import comflower.sagongsa.request.CreateCommentRequest;
 import comflower.sagongsa.request.EditCommentRequest;
-import comflower.sagongsa.entity.Comment;
-import comflower.sagongsa.entity.Post;
-import comflower.sagongsa.error.CommentNotFoundException;
-import comflower.sagongsa.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,37 +17,58 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final PostRepository postRepository;
 
-    public boolean isCommentPresentById(Long commentId) {
-        return commentRepository.findById(commentId).isPresent();
+    public List<Comment> getCommentsByPost(Long postId) {
+        if (postRepository.findById(postId).isEmpty()) {
+            throw new PostNotFoundException(postId);
+        }
+
+        return commentRepository.findByPostId(postId);
     }
 
-    public Comment getComment(Long commentId) {
-        return commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException(commentId));
-    }
+    public Comment createComment(Long authorId, Long postId, CreateCommentRequest request) {
+        if (postRepository.findById(postId).isEmpty()) {
+            throw new PostNotFoundException(postId);
+        }
 
-    public List<Comment> getCommentsByPost(Post post) {
-        return commentRepository.findByPostId(post.getId());
-    }
+        Long parentId = request.getParentId();
+        if (parentId != null && commentRepository.findById(request.getParentId()).isEmpty()) {
+            throw new CommentNotFoundException(parentId);
+        }
 
-    public Comment createComment(Long authorId, Long postId, CreateCommentRequest body) {
         Comment comment = Comment.builder()
                 .postId(postId)
                 .authorId(authorId)
-                .content(body.getContent())
+                .content(request.getContent())
                 .createdAt(LocalDateTime.now())
-                .parentId(body.getParentId())
+                .parentId(request.getParentId())
                 .build();
         return commentRepository.save(comment);
     }
 
-    public Comment editComment(Comment comment, EditCommentRequest body) {
-        comment.setContent(body.getContent());
+    public Comment editComment(Long postId, Long commentId, EditCommentRequest request) {
+        if (postRepository.findById(postId).isEmpty()) {
+            throw new PostNotFoundException(postId);
+        }
+
+        Comment comment = getComment(commentId);
+        comment.setContent(request.getContent());
         comment.setEditedAt(LocalDateTime.now());
+
         return commentRepository.save(comment);
     }
 
-    public void deleteComment(Comment comment) {
+    public void deleteComment(Long postId, Long commentId) {
+        if (postRepository.findById(postId).isEmpty()) {
+            throw new PostNotFoundException(postId);
+        }
+
+        Comment comment = getComment(commentId);
         commentRepository.delete(comment);
+    }
+
+    private Comment getComment(Long commentId) {
+        return commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException(commentId));
     }
 }
